+++ src/Vector.cpp
@@ -7,14 +7,14 @@ Vector::Vector() {
     w = 1;
 }
 
-Vector::Vector(double x, double y, double z) {
+Vector::Vector(float x, float y, float z) {
     this->x = x;
     this->y = y;
     this->z = z;
     this->w = 1;
 }
 
-Vector::Vector(double x, double y, double z, double w) {
+Vector::Vector(float x, float y, float z, float w) {
     this->x = x;
     this->y = y;
     this->z = z;
@@ -29,64 +29,64 @@ Vector::Vector(Vector *v) {
 }
 
 Vector Vector::operator+(Vector v) {
-    double x = this->x + v.x;
-    double y = this->y + v.y;
-    double z = this->z + v.z;
-    double w = this->w + v.w;
+    float x = this->x + v.x;
+    float y = this->y + v.y;
+    float z = this->z + v.z;
+    float w = this->w + v.w;
 
     return Vector(x, y, z, w);
 }
 
 Vector Vector::operator-(Vector v) {
-    double x = this->x - v.x;
-    double y = this->y - v.y;
-    double z = this->z - v.z;
-    double w = this->w - v.w;
+    float x = this->x - v.x;
+    float y = this->y - v.y;
+    float z = this->z - v.z;
+    float w = this->w - v.w;
 
     return Vector(x, y, z, w);
 }
 
-double Vector::operator*(Vector v) {
+float Vector::operator*(Vector v) {
     return this->x * v.x + this->y * v.y + this->z * v.z;
 }
 
-Vector operator*(double f, Vector v) {
+Vector operator*(float f, Vector v) {
     return Vector(v.x * f, v.y * f, v.z * f);
 }
 
 Vector Vector::normalize() {
-    double x = this->x / this->w;
-    double y = this->y / this->w;
-    double z = this->z / this->w;
-    double w = 1;
+    float x = this->x / this->w;
+    float y = this->y / this->w;
+    float z = this->z / this->w;
+    float w = 1;
 
     return Vector(x, y, z, w);
 }
 
 Vector Vector::normalize3() {
-    double norm = sqrt(pow(this->x, 2) + pow(this->y, 2) + pow(this->z, 2));
+    float norm = sqrt(pow(this->x, 2) + pow(this->y, 2) + pow(this->z, 2));
 
-    double x = this->x / norm;
-    double y = this->y / norm;
-    double z = this->z / norm;
+    float x = this->x / norm;
+    float y = this->y / norm;
+    float z = this->z / norm;
 
     return Vector(x, y, z, 1);
 }
 
-Vector Vector::operator*(double s) {
-    double x = this->x * s;
-    double y = this->y * s;
-    double z = this->z * s;
-    double w = this->w * s;
+Vector Vector::operator*(float s) {
+    float x = this->x * s;
+    float y = this->y * s;
+    float z = this->z * s;
+    float w = this->w * s;
 
     return Vector(x, y, z, w);
 }
 
-Vector Vector::operator/(double s) {
-    double x = this->x / s;
-    double y = this->y / s;
-    double z = this->z / s;
-    double w = this->w / s;
+Vector Vector::operator/(float s) {
+    float x = this->x / s;
+    float y = this->y / s;
+    float z = this->z / s;
+    float w = this->w / s;
 
     return Vector(x, y, z, w);
 }
@@ -97,11 +97,11 @@ char* Vector::toString() {
     return str;
 }
 
-double Vector::distance(Vector v) {
+float Vector::distance(Vector v) {
     return sqrt(pow(this->x - v.x, 2) + pow(this->y - v.y, 2) + pow(this->z - v.z, 2));
 }
 
-double Vector::norm() {
+float Vector::norm() {
     return sqrt(pow(this->x, 2) + pow(this->y, 2) + pow(this->z, 2));
 }
 
@@ -114,9 +114,9 @@ bool Vector::operator!=(Vector v) {
 }
 
 Vector Vector::cross(Vector v) {
-    double x = this->y * v.z - this->z * v.y;
-    double y = this->z * v.x - this->x * v.z;
-    double z = this->x * v.y - this->y * v.x;
+    float x = this->y * v.z - this->z * v.y;
+    float y = this->z * v.x - this->x * v.z;
+    float z = this->x * v.y - this->y * v.x;
 
     return Vector(x, y, z);
 }
